requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
openai==1.29.0
Flask-Login==0.6.2